+++ django_spire/notification/app/views/json_views.py
@@ -4,6 +4,7 @@ from typing_extensions import TYPE_CHECKING
 
 from django.http import JsonResponse
 from django.contrib.contenttypes.models import ContentType
+from django.contrib.auth.decorators import login_required
 
 from django_spire.notification.app.models import AppNotification
 from django_spire.history.viewed.models import Viewed
@@ -12,6 +13,7 @@ if TYPE_CHECKING:
     from django.core.handlers.wsgi import WSGIRequest
 
 
+@login_required()
 def check_new_notifications_ajax_view(request: WSGIRequest) -> JsonResponse:
     notifications = (
         AppNotification.objects
@@ -26,6 +28,7 @@ def check_new_notifications_ajax_view(request: WSGIRequest) -> JsonResponse:
     })
 
 
+@login_required()
 def set_notifications_as_viewed_ajax(request: WSGIRequest) -> JsonResponse:
     notification_list = (
         AppNotification.objects