+++ django_spire/notification/app/views/page_views.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+from django.contrib.auth.decorators import login_required
 from typing_extensions import TYPE_CHECKING
 
 from django_spire.contrib.generic_views import portal_views
@@ -11,14 +12,16 @@ if TYPE_CHECKING:
     from django.template.response import TemplateResponse
 
 
+@login_required()
 def app_notification_list_view(request: WSGIRequest) -> TemplateResponse:
     app_notification_list = (
         AppNotification.objects.active()
         .is_sent()
         .annotate_is_viewed_by_user(request.user)
-        .select_related("notification")
+        .select_related('notification')
         .distinct()
         .order_by('-notification__sent_datetime')
+        .ordered_by_priority()
     )
 
     return portal_views.list_view(