+++ django_spire/notification/email/helper.py
@@ -3,22 +3,27 @@ from __future__ import annotations
 from django.conf import settings
 from django.core.mail import EmailMessage
 
+from django_spire.notification.models import Notification
+
 
 class EmailHelper:
-    def __init__(
-        self,
-        to: list | str,
-        cc: list | None = None,
-        bcc: list | None = None,
-        fail_silently: bool = False
-    ):
-        if isinstance(to, str):
-            self.to = [to]
+    def __init__(self, notification: Notification, fail_silently: bool = False):
+        email = notification.email
+        if isinstance(email.to_email_address, str):
+            self.to = [email.to_email_address]
         else:
-            self.to = to
+            self.to = email.to_email_address
 
-        self.cc = cc
-        self.bcc = bcc
+        if isinstance(email.cc, str):
+            self.cc = [email.cc]
+        else:
+            self.cc = email.cc
+
+        if isinstance(email.bcc, str):
+            self.bcc = [email.bcc]
+        else:
+            self.bcc = email.bcc
+
         self.from_email = settings.DEFAULT_FROM_EMAIL
         self.fail_silently = fail_silently
 
@@ -26,24 +31,33 @@ class EmailHelper:
 class SendGridEmailHelper(EmailHelper):
     def __init__(
         self,
-        to: list | str,
-        template_data: dict,
-        template_id: str = settings.SENDGRID_TEMPLATE_ID,
-        cc: list | None = None,
-        bcc: list | None = None,
+        notification: Notification,
         fail_silently: bool = False
     ):
-        super().__init__(to, cc, bcc, fail_silently)
+        super().__init__(notification, fail_silently)
 
-        self.template_id = template_id
-        self.template_data = template_data
+        if notification.email.template_id == '':
+            self.template_id = settings.SENDGRID_TEMPLATE_ID
+
+        else:
+            self.template_id = notification.email.template_id
+
+        self.template_data = {
+            'to': notification.email.to_email_address,
+            'from': settings.DEFAULT_FROM_EMAIL,
+            'subject': notification.title,
+            'body': notification.body,
+            'link': notification.url
+        }
+
+        self.template_data.update(notification.email.context_data)
 
     def send(self) -> None:
         msg = EmailMessage(
             from_email=self.from_email,
             to=self.to,
             cc=self.cc,
-            bcc=self.bcc
+            bcc=self.bcc,
         )
         msg.template_id = self.template_id
         msg.dynamic_template_data = self.template_data