+++ docs/changelog/changelog.md
@@ -1,5 +1,41 @@
 # Changelog
 
+## v0.8.3
+
+### Fixes
+
+- Fixed missing login required decorator on multiple views.
+
+### Features
+
+- Added custom template_id, context_data, bcc and cc to `EmailNotification` model
+
+## v0.8.2
+
+### Fixes
+
+- Fixed missing login required decorator on multiple views.
+
+### Features
+
+- Added `App Notification` list view filtering to include priority.
+
+
+## v0.8.1
+
+### Fixes
+
+- Fixed missing login required decorator on `app_notification_list_view`.
+
+
+## v0.8.0
+
+### Features
+
+- Base service layer and descriptor to extend models.
+- Default service for common functionality (saving an instance).
+
+
 ## v0.7.1
 
 ### Fixes